Try fix build

+++ run_tab2json.py
@@ -4,6 +4,7 @@ import sys
 import os
 import json
 import shutil
+import tempfile
 src_zip = sys.argv[1]
 
 try:
@@ -11,11 +12,16 @@ try:
 except ImportError as e:
     raise RuntimeError("Could not import isatools package")
 print("src_zip = " + src_zip)
-shutil.unpack_archive(filename=src_zip)
+tmpdir = tempfile.mkdtemp()
+print('unpacking to ' + tmpdir)
+shutil.unpack_archive(filename=src_zip, extract_dir=tmpdir, format='zip')
 
-print(os.getcwd(), os.listdir(os.getcwd()))
+if not os.path.exists(tmpdir):
+    print("File path to ISA-Tab files {} does not exist".format(tmpdir))
+    sys.exit(0)
 
-my_json = isatab2json.convert(work_dir='/', validate_first=False, use_new_parser=True)
+print(os.listdir(tmpdir))
+my_json = isatab2json.convert(work_dir=tmpdir, validate_first=False, use_new_parser=True)
 with open("/out.json", 'w') as out_fp:
     json.dump(my_json, out_fp)
     with open("/out.json") as j_fp:
